requirements.txt
groq
sentence-transformers
faiss-cpu
pandas
streamlit
numpy